+++ portal/src/components/FacilityActivation/FacilityActivation.js
@@ -12,7 +12,7 @@ import DetailsCard from "../DetailsCard/DetailsCard";
 function FacilityActivation({
                                 facilities, setFacilities, selectedState, onStateSelected, districtList, selectedDistrict,
                                 setSelectedDistrict, stateList, programs, selectedProgram, setSelectedProgram, facilityType, setFacilityType,
-                                status, setStatus, fetchFacilities, resetFilter
+                                status, setStatus, fetchFacilities, resetFilter, updateFacilityProgramStatus
                             }) {
 
     const [allChecked, setAllChecked] = useState(false);
@@ -85,22 +85,7 @@ function FacilityActivation({
 
     const handleActiveClick = () => {
         setAllChecked(false);
-        if (selectedProgram && selectedFacilities.length > 0) {
-            let updateFacilities = [];
-            selectedFacilities.forEach(facility => {
-                let programs = [{
-                    id: selectedProgram,
-                    status: status !== CONSTANTS.ACTIVE ? CONSTANTS.ACTIVE : CONSTANTS.IN_ACTIVE
-                }];
-                updateFacilities.push({osid: facility.osid, programs})
-            });
-            axiosInstance.current
-                .put(API_URL.FACILITY_API, updateFacilities)
-                .then((res) => {
-                    //registry update in ES happening async, so calling search immediately will not get back actual data
-                    setTimeout(() => fetchFacilities(), 2000);
-                });
-        }
+        updateFacilityProgramStatus(selectedFacilities, oppositeStatus);
     };
 
     return (
@@ -179,6 +164,8 @@ function FacilityActivation({
                     showCard={showCard}
                     setShowCard={setShowCard}
                     data={selectedRow}
+                    status={status}
+                    updateFacilityProgramStatus={updateFacilityProgramStatus}
                 />
             </div>
             <div className="col-sm-3 container">
